+++ src/senaite/storage/browser/workflow/analysisrequest.py
@@ -35,16 +35,3 @@ class WorkflowActionStoreAdapter(RequestContextAware):
         url = "{}/storage_store_samples?uids={}".format(
             api.get_url(self.context), ",".join(uids))
         return self.redirect(redirect_url=url)
-
-
-@implementer(IWorkflowActionUIDsAdapter)
-class WorkflowActionBookOutAdapter(RequestContextAware):
-    """Adapter in charge of Analysis Requests 'book_out' action
-    """
-
-    def __call__(self, action, uids):
-        """Redirects the user to the book out form
-        """
-        url = "{}/storage_book_out_samples?uids={}".format(
-            api.get_url(self.context), ",".join(uids))
-        return self.redirect(redirect_url=url)