+++ src/senaite/storage/permissions.py
@@ -1,14 +1,19 @@
 # -*- coding: utf-8 -*-
 
+# Add Permissions
+AddStorageRootFolder = "senaite.storage: Add Storage Root Folder"
+AddStorageFacility = "senaite.storage: Add Storage Facility"
+AddStorageContainer = "senaite.storage: Add Storage Container"
+AddStorageSamplesContainer = "senaite.storage: Add Storage Samples Container"
+AddStoragePosition = "senaite.storage: Add Storage Position"
+
 # Permissions when context is a sample
 TransitionStoreSample = "senaite.storage: Transition: Store Sample"
 TransitionRecoverSample = "senaite.storage: Transition: Recover Sample"
-TransitionBookOutSample = "senaite.storage: Transition: Book out Sample"
 
 # Permissions when context is a storage samples container
 TransitionAddSamples = "senaite.storage: Transition: Add Samples"
 TransitionRecoverSamples = "senaite.storage: Transition: Recover Samples"
-TransitionBookOutSamples = "senaite.storage: Transition: Book out Samples"
 
 # Permissions when context is a storage content
 TransitionDeactivate = "senaite.storage: Transition: Deactivate"